AU Host:  Ensure discrete parameters are scaled correctly
Previously, the AU hosting code always computed the number of steps in
the parameter range as though the minimum parameter value was 0.

Now, we take the parameter's reported minimum into account when
computing the number of steps. We also use the parameter's range, rather
than its step number, when normalising/denormalising the parameter
value.